fix: function naming and locations
